Refine webhooks
The trigger webhook now watches referenced services and workloads. It
also ignores dry run requests.

The projector sets the default volume file permissions to avoid a
phantom diff reconciliation race. This is a stopgap as we should observe
the diff and track it going forward (similar to what ChildReconciler
does internally).

Signed-off-by: Scott Andrews <scott@andrews.me>