Check RBAC access to target resource before intercepting with a webhook
Admission webhooks are bypass the normal RBAC access controls used by a
controller when communicating with the API Server. We can use a
SelfSubjectAccessReview to verify that the manager's service account has
a permission to a resource before we intercept it with a webhook.

The projector mutating webhook checks for "update" permission. The
trigger webhook checks for "get" permission. If the permission is not
allowed, the resource is not added to the webhook rules. Errors are
treated as denials.

Each resource permission requires an independent SelfSubjectAccessReview
check, which can be chatty, but is more reliable in clusters with
dynamic authorization controllers installed.

The access review check is cached for 5 minutes before being refreshed.

Signed-off-by: Scott Andrews <scott@andrews.me>